docwriter: Use <pre> for code blocks
Code blocks, i.e. |[ ]|, were converted into <code> elements, but <pre>
would be more appropriate.

In addition, add four spaces to the start of each line in a <pre> element
so that Markdown will recognize it as code and not mangle it. As explained
in a FIXME comment, a better solution would be to reimplement
DocstringScanner using Markdown extensions so that such confusion becomes
impossible.